plan for future percentage calculating algorithm

+++ src/main/java/com/example/application/repositories/CardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CardRepository extends JpaRepository<Card, Long> {
-    List<Card> findByDeckId(Long deckId);
+public interface CardRepository extends JpaRepository<Card, Integer> {
+    List<Card> findByDeckId(Integer deckId);
 
 }
